Library Dynamic Hours
This JavaScript I wrote was to solve the issue of having to go into the HTML to update the hours of when the building is open. The old website used an iframe to a Wordpress widget that was on the main La Salle URL. With the new website, they moved away from this since they wanted to have more control over the styling and the feature was cumbersome. The library website is hosted with Springshare and the solution I found was to fetch a CSV file that was stored on our back-end assets page. Since the content ID of the assets do not change when content is modified, it can be downloaded, edited, and then re-uploaded within Springshare's backend. This makes it easier for someone without any HTML knowledge to be able to edit one CSV file to update the hours on the website and not accidently remove any HTML.