Library Fix Proxy Links
This program was used for the purpose of fixing many of the links that were on the website with the old/new proxy attached to the URL of each link. Springshare has a link checker feature to create a report of broken or unreachable links. In summer of 2022, when the library moved to a new authentication method with OpenAthens, many of the links were "hard-coded" to have the old proxy method called dbproxy in the URL instead of using the built in proxy option. Some newer links were also being created by staff with the new OpenAthens proxy "hard-coded" since the global proxy was never changed over from dbproxy until I updated it. This caused the link checker to be flooded with any link using the old dbrpoxy method (over 900 links) and there were new links with the new proxy that would break if there was ever a switch again. With this script, it filters and removes either of the proxies that are "hard-coded" on the URL and instead checks the option to enable the OpenAthens proxy.