Library Link Filter
The purpose of this program was to filter many of the broken links that were found on the library website. The library website is hosted on Springshare and has a built-in link checker tool to keep track of which links are broken or unreachable. The problem was that the link checker was not clear on why each link was down and also contained some false positives. With this script, it takes all of the links from the link checker, tries to get a response from each link, and then writes the link to a file depending on what response was returned. This made it easier to get an idea why each link was marked broken and what to expect from the links. It also ignored and links that currently were not mapped to a page, but still in the assets.