Library Link Check Test
The purpose of this program was to speed up the process of being able to check each of broken links found from the library website. The problem was that testing each link was tedious and could take time to check it with the proxy and then update its status. This automates the process of opening each link, opening the link with the proxy and with the encoded proxy and bringing up the link asset within Springshare (The library's website host).